refactor legacy codes

+++ src/ExcelSheetColumnNumber.java
@@ -1,3 +1,14 @@
+//Given a column title as appear in an Excel sheet, return its corresponding column number.
+//
+//        For example:
+//
+//        A -> 1
+//        B -> 2
+//        C -> 3
+//        ...
+//        Z -> 26
+//        AA -> 27
+//        AB -> 28
 
 public class ExcelSheetColumnNumber {
 
@@ -7,15 +18,15 @@ public class ExcelSheetColumnNumber {
         }
         int sum = 0;
         for (int i = 0; i < s.length(); i++) {
-            int digit = (int) (s.charAt(i) - 'A' + 1);
+            int digit = s.charAt(i) - 'A' + 1;
             sum = sum * 26 + digit;
         }
         return sum;
     }
-    
+
     // unit test
     public static void main(String[] args) {
-    	String s = "AB";
-    	System.out.println(titleToNumber(s)); // 28
+        String s = "AB";
+        System.out.println(titleToNumber(s)); // 28
     }
 }
